remove basket functionality; implement NewFromBundle
